[DRAFT] Configure ACL via static file
This is a draft patch that allows user to configure vFW ACL
via configuration file provided in TC definition. For now,
Yardstick applies the rules to SampleVNF from specified config
file + rules generated by Yardstick. This behaviour may be chaged and
it depends on the provided requirentmets. Only vFW is affected
for now. The `fwd` & `nat` action using hardcodded port 0 for now
and it should be changed later.

Change-Id: I76a630261a982083b628e3985fc3bec14ca495db
Signed-off-by: Mytnyk, Volodymyr <volodymyrx.mytnyk@intel.com>