Update patch set 1
Patch Set 1: Code-Review+1

When looking at the packets Sent/received by generator. These reflect
the numbers for the previous interval. The actual sent/received is not as smooth a graph. This is nothing to do with the dashboard itself  but the NSB software so no reason this can not be accepted

Patch-set: 1
Reviewer: Gerrit User 6751 <6751@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1