Update patch set 20
Patch Set 20:

Hi Volodymyr:

MQ are faster, scalable and as you pointed out in your comment, could be used with external processes. With this architecture we also have event driven classes which can run in parallel to the consumer processes.

By just implementing the consumer or the producer classes in child objects, we immediately have event aware objects capable to send/receive messages and execute actions depending on those messages.

About the implemented messages: start, end and iteration are, initially, the main events a runner class should implement. Although this class is not implementing them, we give to the user this option.

Patch-set: 20
Label: Verified=0