Bugfix: HA kill process recovery has a conflict
It happens in Nokia SUT when running in the Plugfest. The problem happens when
the start_process start to recover the killed process (like nova-api), but the
self-cured mechanism already recovered. And somehow it lead to a conflict and
has problems.
So the recover of the HA attack-recover should be improved to only recover it
when it needs to.

JIRA: YARDSTICK-1222

Change-Id: I66b28198057774fb99bdbd00e984a5b75b9d13b4
Signed-off-by: rexlee8776 <limingjiang@huawei.com>