Create a SampleVNF MQ consumer class
This new class is designed to be used in a ``GenericVNF`` object (VNF
or traffic generator) in a builder pattern. This consumer class will
be instantiated when initializing the ``GenericVNF`` derived object.
The ``GenericVNFConsumer`` endpoint is the ``GenericVNF``derived object
itself.

``GenericVNFEndpoint`` is an interface class to be inherited in a
``GenericVNF`` child object; the VNF object will implement the endpoints
methods.

  NewVNF(GenericVNF, GenericVNFEndpoint):
    def __init__(self, name, vnfd, task_id):
      ...
      self._consumer = vnf_base.GenericVNFConsumer([task_id], self)
      self._consumer.start_rpc_server()

JIRA: YARDSTICK-1293

Change-Id: I1bcd980352099e9ebefdab0c96d51aa46f6a1e54
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>