Place test specific settings in test file and override traffic profile settings when defined
JIRA: YARDSTICK-1298

Allow the test file eg "tc_prox_baremetal_l2fwd-4.yam"
setting to override the values in the traffic_profile file
eg "prox_binsearch.yaml"

This is done by adding a section to "options section" in the test file.
See below :-

  options:
    interface_speed_gbps: 10

    traffic_config:
      tolerated_loss: 0.001
      test_precision: 0.1
      packet_sizes: [64]
      duration: 30
      lower_bound: 0.0
      upper_bound: 100.0

    vnf__0:
      prox_path: /opt/nsb_bin/prox
      prox_config: "configs/handle_l2fwd-4.cfg"
      prox_args:
        "-t": ""

    tg__0:
      prox_path: /opt/nsb_bin/prox
      prox_config: "configs/gen_l2fwd-4.cfg"
      prox_args:
        "-e": ""
        "-t": ""

If the section "traffic_config" doies not exist in the test file, the
these values are taked from the traffic_profile ".yaml" file.

All, some values of none can be defined in the traffic profile section
of options, the missing items are then taken from the traffic profile
.yaml file.

If they do not exist in the yaml file then default values are used.

Change-Id: I6b4e123e865c9dab548f9ad28647f95b53410e8f
Signed-off-by: Daniel Martin Buckley <daniel.m.buckley@intel.com>