Disk Initialization and Custom Workload
Removes the deprecated nowarm and nosdd parameters from the
call to StorPerf.

Adds a call to initialize the disk right after the stack
creation is complete so that the results from the
performance run are valid and not skewed by a disk that
is not operating at a consistent level of performance.

Adds the jobs API v2.0 when the custom parameter 'workloads'
is used instead of workload.

Improves documentation and adds more info about the
workload and workloads parameters.

Co-Authored-By: Ameed Ashour <Ameed.Ashour.Ext@nokia.com>

Change-Id: I17e96028f6a82e2c800c8f73c7be866e1cab6ad8
JIRA: YARDSTICK-1322
Signed-off-by: mbeierl <mark.beierl@dell.com>