Update patch set 2
Patch Set 2:

(2 comments)

Does it make sense to keep this consistent with the fio.yaml? i.e. the only difference should be in the context section?

If so, it might be as easy to add a {% if context = k8s|heat %} block to the fio scenario file.

Patch-set: 2
Label: Code-Review=0