Update patch set 3
Patch Set 3:

> > Same as for the other k8s TC, can it be part of the lmbench.yaml,
 > > and add a second context which can be chosen through task args
 > 
 > I think this makes sense. I also would like to make like tc002 to
 > support both heat and k8s context, and with option to switch. One
 > problem need to fix is that k8s name do not support "_".

In this case, you can add another {% if context ==... %} block around the hostname. 

I thought it was '.' that k8s didn't support, which means that you need a different name to refer to the servers in the different contexts (demeter-k8s vs demeter.demo) 
The issue here is the hostname and the use of separators, right?
If so, you could add a flag the context to allow you to specify the host name separator in heat to be '-' instead of the default '.'.
This would take a little longer to implement than the {% if... %} blocks though.

[1] https://github.com/opnfv/yardstick/blob/master/yardstick/benchmark/contexts/base.py#L47-L59

Patch-set: 3
Label: Code-Review=0