Update patch set 7
Patch Set 7: Code-Review+1

(1 comment)

I'm open to discussing and helping you to get to using Yardstick contexts to help automate the provisioning of your stack and volumes.

I've suggested the Node context in the comments, however, depending on how you're creating your heat stack, you could be in a position to help test the no_setup/no_teardown flags and verify that they work when you use a heat template.

We can discuss this more on the mailing list or over IRC

Patch-set: 7
Label: Code-Review=+1