Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

This patch is way too big to review properly in any reasonable time. 

Definitely, you should not have the vEPC introduced in the same patch as the landslide changes.

I suggest that each change should include just one class and its associated unit tests.
This would be easier to review, 2.5k is not practical to review, as it is much harder to track multiple features and feedback.

Patch-set: 1
Label: Code-Review=-1