Update patch set 1
Patch Set 1:

> These are yaml files and they dont need to be rebased. These needs
 > to be merged once the api's are merged. So request to review based
 > on contents and not the link between implementations.

The link between commits is important, and the order in which these merge is important so we don't end up with invalid testcases that can't be run because the implementation is under review.
Verified-1 prevents the change from merging out-of-sequence, as there is no obvious prerequisite for this patch.

Patch-set: 1