Update patch set 1
Patch Set 1:

(7 comments)

Please make the placeholders more obvious, so that a user can identify what needs to be filled in with a quick glance. The user should intuitively know what value to put into the placeholders as well, without needing to consult the user guide.

Please also rebase this on top of the dependant code changes, so that someone wanting to verify it can pull a series of patches and test it out.
Rebasing this change on top of the code changes also helps people understand what features are related. It would also make it a whole lot easier should this change need to be backported at a later stage.

Patch-set: 1