Update patch set 1
Patch Set 1:

> (1 comment)

It still does not do the same thing. If we would use safe_ip_address, it does accepts ipv6, and make_ip_addr returns an IPVXInterface object. I thought it was the most practical approach to create a new function instead of tweaking the old ones.

Feel free to suggest a better solution :)

Patch-set: 1