Update patch set 1
Patch Set 1:

> > > (1 comment)
 > >
 > > It still does not do the same thing. If we would use
 > > safe_ip_address, it does accepts ipv6, and make_ip_addr returns
 > an
 > > IPVXInterface object. I thought it was the most practical
 > approach
 > > to create a new function instead of tweaking the old ones.
 > >
 > > Feel free to suggest a better solution :)
 > 
 > In this case you can use `version` attribute of IPv4Address or
 > IPv6Address objects (returned by ipaddress.ip_address) to validate
 > if the IP is IPv4 address and raise an Yardstick exception in case
 > it is not.

I know, but then I'd have to either check for errors in every replaces assert statement in PipelineRules, or create a function for it, which seems like unnecessary overhead to me. Then I could have created a a function that uses safe_ip_address and then checks for errors. However, it just seemed more practical to add a new function which does the same thing as safe_ip_address, but skips the IP version check function in ipaddress.py, and does not skip raised errors.

Patch-set: 1