NSB NFVi PROX BNG and vPE losing many packets
JIRA: YARDSTICK-1101

The traffic generated by the generator in bng-4 result in
non-symmetrical use of the cores on the SUT, as a whole
range of IP addresses is skipped.

Update cpe0 and cpe1 in gen files.

In addition, for the routing table (ipv4.lua) for the BNG
should be slighly different: while the BNG is a 4 ports
test, only 2 of the ports are used as outgoing routing
ports (the inet ports). Hence, the routing table should
only use the mac of those two ports. The bad configuration
(using 4 ports) result in an issue when BNG SUT and GEN
are connected through a switch, packets are sent by the
BNG SUT through the INET interfaces, but sometimes with
the MAC of the CPE interface...

Hence, packets are received on the wrong interface on the
Gen site (resulting in some interfaces, the CPE ones,
being overloaded)

Created ipv4_bng.lua and remove mac 0 and 2

Change-Id: I97019f0415d2f58d50ec855d04800eb705f41541
Signed-off-by: Daniel Martin Buckley <daniel.m.buckley@intel.com>