Update patch set 11
Patch Set 11: Code-Review-1

You seem to be making two changes here.
* adding the CD-ROM device
* Updating user access

The first item is not mentioned in the commit message, and I'm clear as to how it's related to the rest of the change.
If it's related, please let me know how, and if it's not, it should be in a separate change.

Patch-set: 11
Reviewer: Gerrit User 2311 <2311@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1