Yardstick manifest for kubernetes
As a user, he/she wants to run Yardstick in a kubernetes scenario.
It's better that Yardstick has a manifest yaml so he/she can apply
it to the kubernetes and start it quickly.
3 deployment:
    - yardstick-server
    - yardstick-influxdb
    - yardstick-grafana
besides the docker, the config of dockers should be automatically done
via the manifest. This includes:
    - yardstick.conf point to influxdb database
    - influxdb with user and database created
    - grafana with all dashboards imported
So it requires to use configMap to config the dockers at the same time.

JIRA: YARDSTICK-1376

Change-Id: Ib1bbc63ec64ccda4dc7276c7fc42c6eb04af92e2
Signed-off-by: rexlee8776 <limingjiang@huawei.com>