Update patch set 6
Patch Set 6: Code-Review+1

> The precision of the tolerance is already defined in the interval.
 > This patch is unneeded.


This patch is adding new status field to the result which is used to select the best throughput value of the iterations. The precision (which is calculated from allowed drop percentage) is used to calculate the correct iteration state.
Without this patch, the user will not be able to show the throughput on the Grafana.

Patch-set: 6
Label: Code-Review=+1