Update patch set 6
Patch Set 6:

As I said, if the calculated drop percentage is in between the given drop percentage range, the test should stop and this iteration is the best one.

There is a flag, "completed", which is given to "run_traffic" to stop any further traffic injection. If any specific field must be added (not needed because the last iteration is the best one), this field must be linked to this "completed" flag.

Patch-set: 6