Update patch set 1
Patch Set 1:

Hi Rex, thanks for working on this. One general question: shouldn't the versions used by Yardstick master be aligned with the upper constraints of OpenStack Queens and Yardstick stable/fraser with Pike? If using the pike constraints for queens, you might use packages which are too old for the master branch.

Patch-set: 1