Update patch set 6
Patch Set 6:

> As I said, if the calculated drop percentage is in between the
 > given drop percentage range, the test should stop and this
 > iteration is the best one.
 > 
 > There is a flag, "completed", which is given to "run_traffic" to
 > stop any further traffic injection. If any specific field must be
 > added (not needed because the last iteration is the best one), this
 > field must be linked to this "completed" flag.

The test shouldn't stop once the calculated drop percentage is in between the given drop percentage range. For example, the first iteration is running with 100% rate and the drop percentage rate is not in the given range. Then, the yardstick will choose next rate 50% for next iteration. With 1/2 of rate the drop percentage likely will be in the given range but the 50% is doesn't mean the BEST throughput value in this case anyway. Addition iterations should be performed to get the best throughput value. Please note, that there is not quarantine that the next chosen value will be in the given range. Thus, the new DB field is introduced for the user to know/select the best value from the throughput values which are in the given range.
Ideally, there should be a "iterations" configuration filed in traffic profile which should indicate TG how many iteration should be performed by RFC2544 similar to what IXIA RFC2544 quick test does but this can be implemented in the future. Now, yardstick uses runner for this purpose.

Patch-set: 6