NSB PROX NFVi Test does not stop after reaching expected test_precision
JIRA: YARDSTICK-1419

When using prox_binserach algorithm, a binary search is
performed, increasing lower bound when step was sucessful
and decreasing upper bound when the step was a failure.

This runs until the test_precision (as specified inthe
traffic profile) is reached. When the test precision
is reached, the test is not completed until the the
runner duration is reached.

As runner duration is usually high (e.g. 1800sec), the
tests take much too long to execute.
This makes it difficult to create test suites.

Change-Id: I6cc503a09fb534a556c61c805e6df4786bb8cc4b
Signed-off-by: Daniel Martin Buckley <daniel.m.buckley@intel.com>