Update patch set 4
Patch Set 4: Code-Review+1

Apologies, I reviewed my comments and hit "Post" before adding a cover message or review-score.

I added tickets to follow up on items, other than that, it seems that everything was addressed, and my questions answered.

I'll give this a +1. Validation is pending, and once it's confirmed that everything is working as expected, I'm happy to +2

Patch-set: 4