Fix NSB PROX L3FWD dropping packets
JIRA: YARDSTICK-1037

Adjusted number of RX/TX descriptors to 2048, instead of default 256,
in order to cope with interrupts during which no packets are received
by the core, which results in the NIC dropping packets when it gets out
of available descriptors to buffer them.
Also adjusted mempool size, because more descriptors need more mbufs.

This change aligns L3FWD test case with other test cases, for which
number of RX/TX descriptors is already set to 2048.

Change-Id: I11378c078398aca799295b619bb3d14dbee04ca3
Signed-off-by: Patrice Buriez <patrice.buriez@intel.com>