[WIP] Fix QnQ issue
Currently if multiple tests are run with a QnQ section defined the QnQ
segment gets appended to the frame. This leads to invalid frames when
the test is run for more than 1 iteration. With this patch the QnQ is
updated if the QnQ segment has already been created.

JIRA: YARDSTICK-1513

Change-Id: I4ba6ef89380ccf1cb8a132ff5bf048935a87f0c9
Signed-off-by: John O Loughlin <john.oloughlin@intel.com>
Signed-off-by: Oleksandr Naumets <oleksandrx.naumets@intel.com>