[docs][nsb][install] Update source for sample pod files
The command for copying the sample pod files was incorrect
in one instance (baremetal config) and inconsistent in
other instances (using ./etc vs <yardstick>/etc vs <repo>/etc)

This updates 13-nsb-installation.rst for consistency.
Paths within the yardstick repo are preface with ``./``,
as initially, the user was instructed to ``cd yardstick``.

JIRA: YARDSTICK-1335
Change-Id: Ia5d5fae29373a177f775ea433c56f38bdeca536e
Signed-off-by: Emma Foley <emma.l.foley@intel.com>