Update patch set 2
Patch Set 2:

(5 comments)

Just minor comments and it is good to go. This is probably the first template in for the prox VNF, so everyone will be copying it to create new one. So, let's try to keep it as a good example as much as possible :)

Thanks.

Patch-set: 2
Label: Verified=0