Add ability to get data from different DBs in influx
In order to get data from Barometer, the influx helper need
to be updated to recognise different DBs. This patch adds
an optional arg to specify which database to use.
By default, the configured dispatcher for Yardstick is
used.
Limitation: The new change hard codes barometer DB name to
be used, and assumes that it is in the same instance of
Influx as the Yardstick dispatcher.
This can be extended later to make it configurable,
however, a new config option would have to be introduced.

JIRA: YARDSTICK-1593
Change-Id: Idee9c3491d6bf7b549e014edd9ff031312cf0ec2
Signed-off-by: Emma Foley <emma.l.foley@intel.com>