Update patch set 5
Patch Set 5: Code-Review-1

(11 comments)

If this is not ready to be reviewed, please note that the convention we use is to put "[WIP]" as the start of the subject line in the commit message.

Please consult the style guidelines [1] in the devguide for the next patchset/when you are ready for review.

[1] https://github.com/opnfv/yardstick/blob/master/docs/testing/developer/devguide/devguide.rst#development-guidelines

Patch-set: 5
Label: Code-Review=-1