Update patch set 1
Patch Set 1: Code-Review+1

This is definitely workaround and it doesn't resolves the root cause. If this is critical issue, I don't see any problems here to get this patch merged since it changes just a socket timeout and in normal operation it should work as it was before.

Patch-set: 1
Label: Code-Review=+1