Leverage on PBR
It fixes requirements and constraints by switching to PBR [1].
It synchronizes requirements with OpenStack stable/stein and then
adds constraints to pip calls.

It also updates flask-restful-swagger to conform with OpenStack
upper-constraints [2] and the pip API calls.

It updates the unit tests verifying a few package versions.

It stops checking unicodecsv in py3 env which is against the
requirements defined in OpenStack.

[1] http://testresults.opnfv.org/functest/dockerslicing/
[2] https://github.com/rantav/flask-restful-swagger/commit/fd418b50723fe1e79f495aa400a8e9f055a5be27

Change-Id: I635bdb9d7411e619cee911b633fa8ad591a393a0
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>