Update patch set 1
Patch Set 1:

Hello 
i have updated this file as Energy is not only a new scenario type but also a deep mind change starting because idea is the capability to consider Energy consumption for each test. Even if it is for each test, it is a new category not yet present inside yardstick. 
It will be next considered for virtual environment as well.
Thanks if you could review my contribution which is part of 3-part contribution. Other is done by me regarding energy test case and the last one is the code pushed by Guannan
KR
Roland.

Patch-set: 1
Label: Verified=0